refactor(utils): reduce re-exports for better tree-shaking
- Remove rulesUtils, varsUtils, pastedImageUtils re-exports from @utils/files
  (consumers already import directly from source modules)
- Remove configConversion re-export from @utils/config
  (convert 2 consumers to direct imports)
- Remove @utils/prompt re-export from @agent/utils
  (no consumers used this path)
- Remove rarely-used internal helpers from xmlUtils barrel:
  convertLatexToMarkdown, convertHtmlToMarkdown, convertWithPandoc,
  extractLatexFromMarkdown, extractLatexBetweenDocumentClass, filterTagsFromText

These changes reduce barrel file complexity and improve tree-shaking by
encouraging direct imports from source modules for rarely-used utilities.
