+++ src/agent/runtime/executeAgent.ts
@@ -70,7 +70,7 @@ import { END_GROUP_STATUS, type EndGroupStatus } from '@logger/messageTypes';
 import { MODEL_CONFIGS } from '@model/ModelRegistry';
 import { updateQueuedFollowUpsUI } from '@progressView/utils/updateQueuedFollowUps';
 import { TaskRunFileService } from '@utils/files';
-import { agentConfigToTaskState } from '@utils/config';
+import { agentConfigToTaskState } from '@utils/config/configConversion';
 import { ensureRunDir } from '@utils/files/taskRunStorage';
 import { bus } from '@eventBus/ProgressEventBus';
 import { getStreamTabId } from '@/logger/streamUtils';
