+++ src/utils/config/index.ts
@@ -1,3 +1,5 @@
 export * from './configUtils';
-export * from './configConversion';
 export * from './constants';
+
+// Note: configConversion is NOT re-exported.
+// Import directly: import { agentConfigToTaskState } from '@utils/config/configConversion';
