+++ src/utils/files/index.ts
@@ -1,11 +1,17 @@
+// Core filesystem abstractions
 export { WorkspaceFS } from './workspaceFS';
 export { AbsoluteFS } from './absoluteFS';
 export { StorageFS, GlobalStorageFS } from './storageFS';
 export { RelativeFS } from './relativeFS';
-export * from './pastedImageUtils';
+
+// Heavily-used utilities (re-exported for convenience)
 export * from './fileMappingUtils';
 export * from './mimeUtils';
 export * from './taskRunStorage';
 export * from './flexibleFS';
-export * from './rulesUtils';
-export * from './varsUtils';
+
+// Note: pastedImageUtils, rulesUtils, and varsUtils are NOT re-exported.
+// Import directly from their source modules:
+//   import { loadTexraRules } from '@utils/files/rulesUtils';
+//   import { setVarFromFile } from '@utils/files/varsUtils';
+//   import { isPastedImage, ... } from '@utils/files/pastedImageUtils';
