+++ src/utils/text/xmlUtils.ts
@@ -3,8 +3,7 @@
  *
  * This module re-exports from focused submodules:
  * - xmlCdata: CDATA section handling
- * - xmlFormatDetection: Format detection (HTML, LaTeX, Markdown)
- * - xmlConversion: Format conversion (Pandoc, Turndown)
+ * - xmlConversion: Format conversion (Pandoc, Turndown) with inlined format detection
  * - xmlExtraction: Content extraction from XML/LaTeX
  *
  * Import directly from the specific module when possible for better tree-shaking.
@@ -17,23 +16,16 @@ export {
   addCdataToTagsMultiple,
 } from './xmlCdata';
 
-// Re-export from conversion module (includes OutputFormat enum)
-export {
-  OutputFormat,
-  convertLatexToMarkdown,
-  convertHtmlToMarkdown,
-  convertWithPandoc,
-  formatContent,
-} from './xmlConversion';
+// Re-export from conversion module
+// Note: convertLatexToMarkdown, convertHtmlToMarkdown, convertWithPandoc are internal helpers
+export { OutputFormat, formatContent } from './xmlConversion';
 
 // Re-export from extraction module
+// Note: extractLatexFromMarkdown, extractLatexBetweenDocumentClass, filterTagsFromText are internal
 export {
   DOCUMENT_NAME_REGEX,
   extractTextFromTag,
-  extractLatexFromMarkdown,
-  extractLatexBetweenDocumentClass,
   extractMultipleTextFromTag,
-  filterTagsFromText,
   extractContentFromXMLbyTag,
   extractContentFromXMLbyTagMultiple,
   extractScratchpad,
